// FrontEnd/ppg-crm/src/app/features/kanban/data/interfaces/stage.interface.ts
import { ProcessDetails } from "./process.interface";


export interface Stage {
    stageId: string;
    projectId: string;
    stageName: string;
    planEndDate: Date;
    processes: ProcessDetails[]
}

// {
//   "projectId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//   "clientId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//   "client": {
//     "clientId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//     "companyName": "string",
//     "contactPerson": "string",
//     "clientEmail": "string",
//     "clientPhone": "string"
//   },
//   "projectName": "string",
//   "description": "string",
//   "status": "NotStarted",
//   "startDate": "2025-08-17T18:33:31.530Z",
//   "endDate": "2025-08-17T18:33:31.530Z",
//   "constructionWorksStart": "2025-08-17T18:33:31.530Z",
//   "budget": 0,
//   "expenses": 0,
//   "progress": 0,
//   "isArchived": true,
//   "stages": [
//     {
//       "stageId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//       "projectId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//       "stageName": "string",
//       "planEndDate": "2025-08-17T18:33:31.530Z",
//       "processes": [
//         {
//           "processId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//           "stageId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//           "processName": "string",
//           "startDate": "2025-08-17T18:33:31.530Z",
//           "planEndDate": "2025-08-17T18:33:31.530Z",
//           "factEndDate": "2025-08-17T18:33:31.530Z",
//           "notes": "string",
//           "notDoneReasons": "string",
//           "problems": "string",
//           "sortOrder": 0,
//           "status": "ToDo",
//           "progress": 0,
//           "totalProcessCost": 0,
//           "tasks": [
//             {
//               "taskId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//               "processId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//               "taskName": "string",
//               "isDone": true
//             }
//           ],
//           "responsibleUsers": [
//             {
//               "userId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//               "userName": "string",
//               "passwordHash": "string",
//               "firstName": "string",
//               "lastName": "string",
//               "email": "string",
//               "phone": "string",
//               "role": "Admin",
//               "salary": 0
//             }
//           ]
//         }
//       ]
//     }
//   ]
// }